feat: add recreate deployment strategy
Recreate means completely stopping your old version of an application and then starting a new version from scratch. It’s like tearing down an old building and building a new one in its place. During this process, the application will be temporarily unavailable while the old version is replaced with the new one.